Error-checking and Windows fixes
* Re-use the exception hook for the error-handling for the EngineWrapper's thread.
* Fix some Pathlib issues on Windows.
